# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AlecRae API
  description: |
    The AlecRae API provides programmatic access to the AlecRae email
    infrastructure platform. Send transactional and bulk emails, manage
    domains, configure webhooks, and query analytics.
  version: 0.1.0
  contact:
    name: AlecRae Support
    url: https://alecrae.com/support
    email: support@alecrae.com
  license:
    name: Proprietary
    url: https://alecrae.dev/terms
servers:
  - url: https://api.alecrae.com
    description: Production
  - url: https://api.staging.alecrae.com
    description: Staging
  - url: http://localhost:3001
    description: Local development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /v1/webhooks:
    post:
      operationId: createWebhook
      summary: Register a webhook
      description: |
        Create a webhook endpoint to receive real-time event notifications.
        The platform signs payloads with HMAC-SHA256 using the provided secret.
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookInput"
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/WebhookRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listWebhooks
      summary: List webhooks
      tags: [Webhooks]
      responses:
        "200":
          description: List of webhooks
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/webhooks/{id}:
    get:
      operationId: getWebhook
      summary: Get webhook details
      tags: [Webhooks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Webhook details
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteWebhook
      summary: Delete a webhook
      tags: [Webhooks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Webhook deleted
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Pass a Bearer token in the Authorization header.
        Example: `Authorization: Bearer em_live_abc123...`
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Pass an API key in the X-API-Key header.
        Example: `X-API-Key: em_live_abc123...`
  headers:
...
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current window
      schema:
        type: integer
        example: 1000
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window
      schema:
        type: integer
        example: 997
    X-RateLimit-Reset:
      description: Unix timestamp (seconds) when the rate limit window resets
      schema:
        type: integer
        example: 1712150400
...
  schemas:
...
    CreateWebhookInput:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
          format: uri
          description: The HTTPS URL to receive webhook events
          example: https://example.com/webhooks/alecrae
        events:
          type: array
          items:
            type: string
            enum:
              - message.sent
              - message.delivered
              - message.bounced
              - message.deferred
              - message.dropped
              - message.complained
              - message.opened
              - message.clicked
              - domain.verified
              - domain.failed
              - contact.subscribed
              - contact.unsubscribed
          minItems: 1
          description: Event types to subscribe to
        secret:
          type: string
          description: HMAC signing secret for payload verification
        description:
          type: string
          maxLength: 256
        active:
          type: boolean
          default: true
    WebhookRecord:
      type: object
      required: [id, url, events, active, createdAt, updatedAt]
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        secret:
          type: string
          description: Redacted in responses
        description:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, message, code]
          properties:
            type:
              type: string
              description: Error category
              example: validation_error
            message:
              type: string
              description: Human-readable error message
              example: "Invalid email address in 'to' field"
            code:
              type: string
              description: Machine-readable error code
              example: invalid_recipient
            context:
              type: object
              additionalProperties: true
              description: Additional error context
  responses:
...
    Unauthorized:
      description: Authentication credentials are missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: The request failed validation (invalid email, domain, etc.)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
        X-RateLimit-Limit:
          $ref: "#/components/headers/X-RateLimit-Limit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/X-RateLimit-Remaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/X-RateLimit-Reset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: An unexpected server error occurred
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"